--- crates/libs/metadata/src/column.rs
@@ -1,11 +1,0 @@
-#[derive(Default)]
-pub struct Column {
-    pub offset: usize,
-    pub width: usize,
-}
-
-impl Column {
-    pub fn new(offset: usize, width: usize) -> Self {
-        Self { offset, width }
-    }
-}
