+++ crates/libs/metadata/src/file.rs
@@ -5,7 +5,7 @@ pub struct File {
     pub bytes: Vec<u8>,
     pub strings: usize,
     pub blobs: usize,
-    pub tables: [Table; 17],
+    tables: [Table; 17],
 }
 
 impl std::fmt::Debug for File {
@@ -579,6 +579,10 @@ impl File {
         first
     }
 
+    pub fn table_len(&self, table: usize) -> usize {
+        self.tables[table].len
+    }
+
     pub fn table<R: AsRow>(&'static self) -> RowIterator<R> {
         RowIterator::new(self, 0..self.tables[R::TABLE].len)
     }
@@ -654,5 +658,63 @@ impl View for [u8] {
         } else {
             None
         }
+    }
+}
+
+#[derive(Default)]
+struct Table {
+    offset: usize,
+    len: usize,
+    width: usize,
+    columns: [Column; 6],
+}
+
+impl Table {
+    fn index_width(&self) -> usize {
+        if self.len < (1 << 16) {
+            2
+        } else {
+            4
+        }
+    }
+
+    fn set_columns(&mut self, a: usize, b: usize, c: usize, d: usize, e: usize, f: usize) {
+        self.width = a + b + c + d + e + f;
+        self.columns[0] = Column::new(0, a);
+        if b != 0 {
+            self.columns[1] = Column::new(a, b);
+        }
+        if c != 0 {
+            self.columns[2] = Column::new(a + b, c);
+        }
+        if d != 0 {
+            self.columns[3] = Column::new(a + b + c, d);
+        }
+        if e != 0 {
+            self.columns[4] = Column::new(a + b + c + d, e);
+        }
+        if f != 0 {
+            self.columns[5] = Column::new(a + b + c + d + e, f);
+        }
+    }
+
+    fn set_data(&mut self, offset: &mut usize) {
+        if self.len != 0 {
+            let next = *offset + self.len * self.width;
+            self.offset = *offset;
+            *offset = next;
+        }
+    }
+}
+
+#[derive(Default)]
+struct Column {
+    offset: usize,
+    width: usize,
+}
+
+impl Column {
+    fn new(offset: usize, width: usize) -> Self {
+        Self { offset, width }
     }
 }
