+++ crates/libs/metadata/src/lib.rs
@@ -8,12 +8,10 @@ mod attributes;
 mod bindings;
 mod blob;
 mod codes;
-mod column;
 mod file;
 mod filter;
 mod reader;
 mod row;
-mod table;
 mod tables;
 mod r#type;
 mod type_name;
@@ -22,13 +20,11 @@ pub use attributes::*;
 pub use bindings::*;
 pub use blob::*;
 pub use codes::*;
-use column::*;
 pub use file::*;
 use filter::*;
 pub use r#type::*;
 pub use reader::*;
 pub use row::*;
-use table::*;
 pub use tables::*;
 pub use type_name::*;
